+++ src/iotools.cpp
@@ -1,7 +1,5 @@
 #ifndef HPCOMBI_IOTOOLS_HPP_INCLUDED
 #include <iotools.hpp>
-#include <sstream>
-#include <string>
 
 void write_hash(int block_size,  const int size, const int nb_vect, uint64_t* hashed, double time, std::string file_name){
 	file_name = "examples/outData/" + file_name + ".data";
@@ -23,7 +21,7 @@ void write_renner(const int size, const int8_t nb_gen, const int count, double t
 }
 
 
-void readRenner(std::string file_name, uint32_t** gen, int* size, int8_t* nb_gen){
+void readRenner(std::string file_name, uint64_t** gen, int* size, int8_t* nb_gen){
 	file_name = "../data/" + file_name;
 	std::ifstream infile(file_name);
 	std::string line;
@@ -41,7 +39,7 @@ void readRenner(std::string file_name, uint32_t** gen, int* size, int8_t* nb_gen
 		exit(1);
 	}
 	
-	(*gen) = (uint32_t*)malloc((*size)*(*nb_gen) * sizeof(uint32_t));
+	(*gen) = (uint64_t*)malloc((*size)*(*nb_gen) * sizeof(uint64_t));
 	std::cout << "Reading file " << file_name << ", size : " << *size << ", nb_gen : " << (int)*nb_gen << std::endl;;
 		
 	for(int i=0; i<*nb_gen; i++){
